+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 - New `gcp_pubsub` input and outputs.
 - New `log` processor.
+- New `lambda` processor.
 
 ## 0.32.0 - 2018-09-18